Update Readme example to use Client-ID header
Previously, the API was promoting the use of the client_id GET parameter in requests. This is functional for all requests, but ```/streams```, which is a unique query parameter for that endpoint. In order to mitigate headaches for newcomers, I have updated the example to reflect a standard request with the Client-ID HTTP header as the preferred method. This is the recommended approach lower in the README under *Rate Limits*.